Put commands into cogs
